Update cors.rst for .NET Core 3.1
I made this same update previously (see https://github.com/IdentityServer/IdentityServer4/pull/4404/files), but it didn't get accepted because I didn't provide the requested justification for my change.  I'll try to provide a good reason here:

I believe that the existing documentation doesn't work, at least for .NET Core 3.1.   This is because an instance of _loggerFactory is no longer available in Startup.cs->ConfigureServices().  Because of this, this  code in the existing documentation breaks:
       var cors = new DefaultCorsPolicyService(_loggerFactory.CreateLogger<DefaultCorsPolicyService>())